ocr-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pillow==10.4.0
pdf2image==1.17.0
paddleocr==2.7.3
paddlepaddle==2.5.2
